chore: updates to work with al2023 (#5)
* updates packages installed on the system

* updates

* updates to use ansible 8.4 with core > 2.13.0 for better Amazon linux support

* updates to use systemd for timer management

* updates vars

* removes geerlingguy.pip role as a dependency

* updates playbook to remove excess

+++ files/amazon-stop-if-inactive.sh
@@ -3,14 +3,13 @@ set -euo pipefail
 IFS=$'\n\t'
 
 # based on AWS Cloud9
-# shutdown ec2 if no ssh connections after 1/2 hour
-#  Add to cron to run every minute
+# shutdown ec2 if no ssh connections after N minutes
 #  User needs sudo permission to run 'shutdown'
 #
 
 set -euo pipefail
 
-SHUTDOWN_TIMEOUT=30     # minutes
+SHUTDOWN_TIMEOUT=60     # minutes
 if ! [[ $SHUTDOWN_TIMEOUT =~ ^[0-9]*$ ]]; then
     echo "shutdown timeout is invalid"
     exit 1
@@ -23,8 +22,25 @@ debug(){
 }
 
 is_shutting_down() {
-    shutdown_status=$(systemctl status systemd-shutdownd | grep -i 'status:' | tr -d '"' | awk '{print $2}')
-    [[ $shutdown_status == Shut* ]]
+    is_shutting_down_system_d &> /dev/null || is_shutting_down_init_d &> /dev/null
+}
+
+is_shutting_down_system_d() {
+    local TIMEOUT
+    TIMEOUT=$(busctl get-property org.freedesktop.login1 /org/freedesktop/login1 org.freedesktop.login1.Manager ScheduledShutdown)
+    if [ "$?" -ne "0" ]; then
+        return 1
+    fi
+    if (( "$(echo $TIMEOUT | awk '{print $3}')" < 0 )); then
+        return 1
+    else
+        # return 0 when shutting down
+        return 0
+    fi
+}
+
+is_shutting_down_init_d() {
+    pgrep shutdown
 }
 
 is_vfs_connected() {
@@ -32,25 +48,24 @@ is_vfs_connected() {
 }
 
 is_ssh_connected() {
-    netstat -tpna | grep 'ESTABLISHED.*sshd' >> /dev/null;
+    ss -tpna | grep 'ESTAB.*sshd' >> /dev/null;
 }
-
 
 isc=$(is_ssh_connected && echo "true" || echo "false")
 if [[ "$isc" == true ]]; then
-    debug "cron - ssh connected. DO NOT schedule a system shut down $0 | $isc"
+    debug "auto-shutdown - ssh connected. DO NOT schedule a system shut down $0 | $isc"
 else
-    debug "cron - ssh not connected. Schedule a system shut down $0 | $isc"
+    debug "auto-shutdown - ssh not connected. Schedule a system shut down $0 | $isc"
 fi
 
 if is_shutting_down; then
-    if is_vfs_connected || [[ "$isc" == true ]]; then
-        debug "cron - cancelling shutdown."
+    if [[ ! $SHUTDOWN_TIMEOUT =~ ^[0-9]+$ ]] || is_vfs_connected || [[ "$isc" == true ]]; then
+        debug "auto-shutdown - cancelling shutdown."
         sudo shutdown -c
     fi
 else
-    if is_vfs_connected || [[ "$isc" == false ]]; then
-        debug "cron - initiating shutdown!!"
+    if [[ $SHUTDOWN_TIMEOUT =~ ^[0-9]+$ ]] && ! is_vfs_connected && [[ "$isc" == false ]]; then
+        debug "auto-shutdown - initiating shutdown!!"
         sudo shutdown -h $SHUTDOWN_TIMEOUT
     fi
 fi
